fix removed m5-d1 timeframes from analize and train

+++ Apache/ApacheAnalizer.py
@@ -20,13 +20,16 @@ import joblib
 
 
 
+
+
+
 class ApacheAnalizer(DataAnalizer):
 
     ############ CONFIG SECTION #############
-    min_records_for_train = 1000 # минимальное количество записей, необходимых для тренировки нс
+    min_for_train = 1000 # минимальное количество записей, необходимых для тренировки нс
     samples_dir = './samples/Apache/'
     model_dir = './models/Apache/'
-    known_classes_count = 0
+    classes_count = 0
     max_urlparams_len = 100
     ############ CONFIG SECTION #############
 
@@ -49,49 +52,50 @@ class ApacheAnalizer(DataAnalizer):
     samples = pd.DataFrame()
 
     def __init__(self, min_for_train) -> None:
-        self.min_records_for_train = min_for_train
-        self.load()
-        self.samples = pd.DataFrame()
-
-
-
-    def save_all_samples(self):
-
-        log('Сохраняем образцы', lvl='info')
-        print(self.samples)
-
-
-        # сохраняем образцы в файлы 
-        grouped = self.samples.groupby('y')
-        for group_name, group_df in grouped:
-            filename = f"{group_name}.csv"  # Имя файла на основе значения y
-            group_df.to_csv(self.samples_dir + filename, sep='\t', index=False, mode='w')
-        log('Образцы сохранены')
-
-    def save_all_models(self):
-        # сохраняем модель в файлы
-        self.model.save(self.model_dir + 'model.h5')
-
-        joblib.dump(self.encoder_methods, self.model_dir + 'encoder_methods.pkl')
-        joblib.dump(self.encoder_y, self.model_dir + 'encoder_y.pkl')
-
-        joblib.dump(self.scaler_count_codes, self.model_dir + 'scaler_count_codes.pkl')
-        joblib.dump(self.scaler_coutn_requests, self.model_dir + 'scaler_coutn_requests.pkl')
-        joblib.dump(self.scaler_trafic_volume, self.model_dir + 'scaler_trafic_volume.pkl')
-        joblib.dump(self.scaler_request_len, self.model_dir + 'scaler_request_len.pkl')
-        joblib.dump(self.scaler_request_code, self.model_dir + 'scaler_request_code.pkl')
-
-        joblib.dump(self.url_param_tokenizer, self.model_dir + 'url_param_tokenizer.pkl')
+        self.min_for_train = min_for_train
+        self.train_samples = pd.DataFrame()
+        self.new_samples = pd.DataFrame()
 
 
     # при завершении работы
     def store(self):
-        self.save_all_samples()
-        self.save_all_models()
-        
-     
-    def load_all_samples(self):
 
+        log('Saving samples and model...', lvl='info')
+        print(self.new_samples)
+            
+        # сохраняем образцы в файлы 
+        grouped = self.new_samples.groupby('y')
+        for group_name, group_df in grouped:
+            filename = f"{group_name}.csv"  # Имя файла на основе значения y
+            if os.path.exists(self.samples_dir + filename):
+                group_df.to_csv(self.samples_dir + filename, sep='\t', index=False, mode='a', header=False)
+            else:
+                group_df.to_csv(self.samples_dir + filename, sep='\t', index=False, mode='w')
+
+        log('Образцы сохранены')
+
+
+        # сохраняем модель в файлы
+
+        try:
+            self.model.save(self.model_dir + 'model.h5')
+
+            joblib.dump(self.encoder_methods, self.model_dir + 'encoder_methods.pkl')
+            joblib.dump(self.encoder_y, self.model_dir + 'encoder_y.pkl')
+
+            joblib.dump(self.scaler_count_codes, self.model_dir + 'scaler_count_codes.pkl')
+            joblib.dump(self.scaler_coutn_requests, self.model_dir + 'scaler_coutn_requests.pkl')
+            joblib.dump(self.scaler_trafic_volume, self.model_dir + 'scaler_trafic_volume.pkl')
+            joblib.dump(self.scaler_request_len, self.model_dir + 'scaler_request_len.pkl')
+            joblib.dump(self.scaler_request_code, self.model_dir + 'scaler_request_code.pkl')
+
+            joblib.dump(self.url_param_tokenizer, self.model_dir + 'url_param_tokenizer.pkl')
+        except:
+            log('Unable save empty model', lvl='error')
+        
+    # загружаем модель и образцы с диска если доступны
+    def load(self):
+        
         files = []
         log('Reading samples from', self.samples_dir, lvl='info')
         if os.path.isdir(self.samples_dir):
@@ -105,11 +109,12 @@ class ApacheAnalizer(DataAnalizer):
         for file in files:
             log('Reading file: ', file, '...', lvl='info')
             df = pd.read_csv(file, sep='\t')
-            self.samples = pd.concat([self.samples, df], ignore_index=True)
+            self.train_samples = pd.concat([self.train_samples, df], ignore_index=True)
 
+        self.classes_count = self.train_samples['y'].nunique()
         log('Reading samples completed', lvl='info')
 
-    def load_all_models(self):
+
         log('Loading model from', self.model_dir, lvl='info')
         files = os.listdir(self.model_dir)
         log('Files in dir: ', files, lvl='info')
@@ -132,32 +137,28 @@ class ApacheAnalizer(DataAnalizer):
         self.model = load_model(self.model_dir + 'model.h5')
 
         log('Model loaded successfully', lvl='info')
-    
-    # загружаем модель и образцы с диска если доступны
-    def load(self):
-        self.load_all_models()
-        self.load_all_samples()
 
 
     # готовим выборку для загрузки в нс
     def prepare_train_data(self, df):
         df['y'] = self.encoder_y.fit_transform(df['y'])
-        self.known_classes_count = len(self.encoder_y.classes_)
+        self.classes_count = len(self.encoder_y.classes_)
 
         df['method'] = self.encoder_methods.fit_transform(df['method'])
 
         code_column = 'code'
         len_column = 'len'
 
-        for column in ['rcm1', 'rcm5', 'rcm30', 'rch3', 'rcd']:
+        for column in ['rcm1']: #, 'rcm5', 'rcm30', 'rch3', 'rcd']:
             df[column] = self.scaler_coutn_requests.fit_transform(df[[column]])
 
-        for column in ['c200m1', 'c200m5', 'c200m30', 'c200h3', 'c200d', 
-                        'c404m1', 'c404m5', 'c404m30', 'c404h3', 'c404d', 
-                        'c403m1', 'c403m5', 'c403m30', 'c403h3', 'c403d']:
+        for column in ['c200m1', # 'c200m5', 'c200m30', 'c200h3', 'c200d', 
+                        'c404m1', # 'c404m5', 'c404m30', 'c404h3', 'c404d', 
+                        'c403m1']:  #, 'c403m5', 'c403m30', 'c403h3', 'c403d']: 
+
             df[column] = self.scaler_count_codes.fit_transform(df[[column]])
 
-        for column in ['volm1', 'volm5', 'volm30', 'volh3', 'vold']:
+        for column in ['volm1']: #, 'volm5', 'volm30', 'volh3', 'vold']:
             df[column] = self.scaler_trafic_volume.fit_transform(df[[column]])
 
         df['len'] = self.scaler_request_len.fit_transform(df[[len_column]])
@@ -185,7 +186,7 @@ class ApacheAnalizer(DataAnalizer):
     # при каждом вызове модель перекомпилируется
     def train_model(self):
 
-        samples_copy = self.samples.copy()
+        samples_copy = self.train_samples.copy()
         X_train, X_test, y_train, y_test = self.prepare_train_data(samples_copy)
 
         y_train = to_categorical(y_train)
@@ -198,7 +199,7 @@ class ApacheAnalizer(DataAnalizer):
         
         self.model.add(LSTM(128, dropout=0.2, recurrent_dropout=0.2))
 
-        self.model.add(Dense(self.known_classes_count, activation='softmax'))
+        self.model.add(Dense(self.classes_count, activation='softmax'))
         self.model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
 
         res = self.model.fit(X_train, y_train, epochs=1, batch_size=32)
@@ -223,36 +224,29 @@ class ApacheAnalizer(DataAnalizer):
         df = pd.DataFrame(records)
         df['y'] = label
 
-
-        # загрузить все с диска
-        self.load()
-
-        # если данных не было на диске то добавляем только что полученные данные в samples
-        if self.samples.empty:
-            self.samples = df
-
-        # а если были, то добавляем новые к существующим
-        else:   
-            self.samples = pd.concat([self.samples, df], ignore_index=True)
-
-            # если количество классов больше 1
-            classes = self.samples['y'].unique()
-            if len(classes) > 1:
-                # готовим отчет с количеством данных каждого из класссов
-                grouped = self.samples.groupby('y')
-                report['classes_count'] = {}
-                for group_name, group_df in grouped:
-                    report['classes_count'][group_name] = group_df.shape[0]
-                    report['accuracy'] = self.train_model()
-
-            else:
-                report['error'] = 'Too less classes for train model'
-                report['more'] = 'Existing classes: ' + ', '.join(classes)
+        self.new_samples = pd.concat([self.new_samples, df], ignore_index=True)
 
 
-        # сохраняем все образцы на диск, чистим samples
-        self.store()
-        self.samples = pd.DataFrame()
+        if self.new_samples.shape[0] >= self.min_for_train:
+            # набрали достаточное количество для тренировки
+            # новые образцы сохр в файлы
+            # прочитьать все образцы из файлов
+            # если классов < 2, ждем новые образцы другого класса
+
+            self.store()
+            self.train_samples = pd.DataFrame()
+            self.new_samples = pd.DataFrame()
+            self.classes_count = 0
+            self.load()
+
+
+            if self.classes_count < 2:
+                log('Classes count < 2. No train', lvl='error')
+                report['error'] = 'Classes count < 2'
+                return report
+            
+            report['accuracy'] = self.train_model()
+
 
         return report
